floatをdoubleに変更（修正漏れ）

+++ src/cylinder.c
@@ -6,7 +6,7 @@
 /*   By: tterao <tterao@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 14:15:30 by ryhara            #+#    #+#             */
-/*   Updated: 2023/10/11 16:13:28 by tterao           ###   ########.fr       */
+/*   Updated: 2023/10/12 14:36:03 by tterao           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,12 +49,12 @@ double	hit_cylinder(t_vector3d ray, t_vector3d start_pos,
 					cylinder->direction)));
 	t.c = vector3d_mag_sq(vector3d_cross(
 				vector3d_sub(start_pos, cylinder->coordinate),
-				cylinder->direction)) - powf(cylinder->radius, 2.0);
-	t.d = powf(t.b, 2.0) - (4.0 * t.a * t.c);
+				cylinder->direction)) - pow(cylinder->radius, 2.0);
+	t.d = pow(t.b, 2.0) - (4.0 * t.a * t.c);
 	if (t.d == 0)
 		t.t = (-(t.b) / (2.0 * t.a));
-	t.t1 = (-t.b + sqrtf(powf(t.b, 2.0) - (4.0 * t.a * t.c))) / (2.0 * t.a);
-	t.t2 = (-t.b - sqrtf(powf(t.b, 2.0) - (4.0 * t.a * t.c))) / (2.0 * t.a);
+	t.t1 = (-t.b + sqrt(pow(t.b, 2.0) - (4.0 * t.a * t.c))) / (2.0 * t.a);
+	t.t2 = (-t.b - sqrt(pow(t.b, 2.0) - (4.0 * t.a * t.c))) / (2.0 * t.a);
 	ray_cylinder = get_cylinder_ray(start_pos, cylinder, t, ray);
 	return (_calc_t(cylinder, ray_cylinder, t));
 }
